+++ CHANGES.md
@@ -1,13 +1,16 @@
 ## Current
 
 - Install EPEL directly from CentOS extras repo
-- Bump FreePBX to 12 (CentOS 6)
+- Bump FreePBX to version 12 on CentOS 6
 - Change the way to detect the OS
+- Zimbra for CentOS 6
+- Switch LEMP PHP 5.4 to archive repo and update phpMyAdmin
+- Few enhancements in nginx and monit configurations
 
 ## 2.0.0 - 2015/08/19
 
 - Major cleanup in CentOS 6 post-installer and several scripts
-- Apache removed from the LAMP build, only LEMP anymore
+- Apache stack removed (LAMP), only LEMP is supported
 - Add OPcache module for LEMP
 - Remove built-in PHP/MySQL option. Only packages from IUS including PHP53 as archive
 - Add RapidSSL 256bits CA bundles (useful for nginx-more and most of my customers!)